fix: fix function 2.2

+++ src/models/Ingredient.java
@@ -5,14 +5,14 @@ import java.io.Serializable;
 public class Ingredient implements Serializable {
     private String code;
     private String name;
-    private int quantity;
+    private double quantity;
     private String unit;
     private double price; //price per unit
 
     public Ingredient(){
     }
 
-    public Ingredient(String code, String name, int quantity, String unit, double price) {
+    public Ingredient(String code, String name, double quantity, String unit, double price) {
         this.code = code.toUpperCase();
         this.name = name.toLowerCase();
         this.quantity = quantity;
@@ -32,11 +32,11 @@ public class Ingredient implements Serializable {
         this.name = name;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
 
@@ -55,12 +55,15 @@ public class Ingredient implements Serializable {
     public void setPrice(double price) {
         this.price = price;
     }
+
+    //quantity round to 2 decimal
+    //price round to 0 decimal
     public void showIngredient(){
-        System.out.printf("| %-5s | %-15s | %10d | %5s | %15.0f |\n",code,name,quantity,unit,price);
+        System.out.printf("| %-5s | %-15s | %10.2f | %5s | %15.0f |\n",code,name,quantity,unit,price);
     }
 
     @Override
     public String toString() {
-        return String.format("| %-5s | %-15s | %10d | %5s | %15.0f |\n", code, name, quantity, unit, price);
+        return String.format("| %-5s | %-15s | %10.2f | %5s | %15.0f |\n", code, name, quantity, unit, price);
     }
 }
